fix(s3): address PR feedback
