+++ src/storage/protocols/s3/index.ts
@@ -1,1 +1,2 @@
+export * from './policy'
 export * from './signature-v4'
